JobsStation/requirements.txt
#di terminal, kalian jalankan dengan cara
#pip install -qU -r requirements.txt

langchain
langchain-openai
langchain-community
langchain-core
langgraph
langfuse
python-dotenv